+++ package.json
@@ -8,10 +8,10 @@
     "packages/*"
   ],
   "resolutions": {
-    "@polkadot/api": "^1.6.0-beta.7",
-    "@polkadot/api-contract": "^1.6.0-beta.7",
+    "@polkadot/api": "^1.6.0-beta.9",
+    "@polkadot/api-contract": "^1.6.0-beta.9",
     "@polkadot/keyring": "^2.6.1",
-    "@polkadot/types": "^1.6.0-beta.7",
+    "@polkadot/types": "^1.6.0-beta.9",
     "@polkadot/util": "^2.6.1",
     "@polkadot/util-crypto": "^2.6.1",
     "babel-core": "^7.0.0-bridge.0",
@@ -30,7 +30,7 @@
     "postinstall": "polkadot-dev-yarn-only",
     "test": "polkadot-dev-run-test packages/page-claims/src",
     "vanitygen": "node packages/page-accounts/scripts/vanitygen.js",
-    "start": "cd packages/apps && webpack --config webpack.config.js"
+    "start": "yarn clean && cd packages/apps && webpack --config webpack.config.js"
   },
   "devDependencies": {
     "@babel/core": "^7.8.7",
