+++ packages/page-contracts/package.json
@@ -12,6 +12,6 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@babel/runtime": "^7.8.7",
-    "@polkadot/api-contract": "^1.6.0-beta.7"
+    "@polkadot/api-contract": "^1.6.0-beta.9"
   }
 }
